Refactor: removed unused jsx from pages folder and moved to components folder, added discord in navbar
