+++ components/footer.jsx
@@ -1,0 +1,13 @@
+export default function Footer() {
+  return (
+    <footer className="flex h-12 w-full items-center justify-center border-t">
+      <p>
+        Developed By{' '}
+        <span className="cursor-pointer underline hover:no-underline">
+          {' '}
+          <a href="https://logan1x.github.io">Logan1x</a>{' '}
+        </span>
+      </p>
+    </footer>
+  )
+}
